Real Time Bus Animation
Receives JSON bus data from the Metro Boston Transit Authority (MBTA) and populates a MapBox API with a custom style overlay showing logos for Harvard and MIT.  The api is repopulated every 15 seconds and a zoom control was added as well.

To run save the files into a local folder and open the index.html file.

Used a personal API token so please use your own.

Would eventually like to change the bus graphic to change at a stop.